+++ app/tests/test_solarman.py
@@ -949,7 +949,7 @@ def test_device_rsp(ConfigTsunInv1, DeviceRspMsg):
     assert m.data_len == 0x0a
     assert m._recv_buffer==b''
     assert m._send_buffer==b''
-    assert m._forward_buffer==DeviceRspMsg
+    assert m._forward_buffer==b''
     assert m.db.stat['proxy']['Invalid_Msg_Format'] == 0
     m.close()
 
@@ -967,7 +967,7 @@ def test_inverter_rsp(ConfigTsunInv1, InverterRspMsg):
     assert m.data_len == 0x0a
     assert m._recv_buffer==b''
     assert m._send_buffer==b''
-    assert m._forward_buffer==InverterRspMsg
+    assert m._forward_buffer==b''
     assert m.db.stat['proxy']['Invalid_Msg_Format'] == 0
     m.close()
 
@@ -1003,7 +1003,7 @@ def test_heartbeat_rsp(ConfigTsunInv1, HeartbeatRspMsg):
     assert m.data_len == 0x0a
     assert m._recv_buffer==b''
     assert m._send_buffer==b''
-    assert m._forward_buffer==HeartbeatRspMsg
+    assert m._forward_buffer==b''
     assert m.db.stat['proxy']['Invalid_Msg_Format'] == 0
     m.close()
 
@@ -1045,7 +1045,7 @@ def test_sync_start_rsp(ConfigTsunInv1, SyncStartRspMsg):
     assert m.data_len == 0x0a
     assert m._recv_buffer==b''
     assert m._send_buffer==b''
-    assert m._forward_buffer==SyncStartRspMsg
+    assert m._forward_buffer==b''
     assert m.db.stat['proxy']['Invalid_Msg_Format'] == 0
     m.close()
 
@@ -1081,7 +1081,7 @@ def test_sync_end_rsp(ConfigTsunInv1, SyncEndRspMsg):
     assert m.data_len == 0x0a
     assert m._recv_buffer==b''
     assert m._send_buffer==b''
-    assert m._forward_buffer==SyncEndRspMsg
+    assert m._forward_buffer==b''
     assert m.db.stat['proxy']['Invalid_Msg_Format'] == 0
     m.close()